claude/turtle-display-key-f46d36d1 (0677844ff, session f46d36d1) has run on the workbench since 2026-09-18 but is not on master; this branch keeps it in the dcore build after the core sync. Switch the checkout back to master once that fix is merged there.

+++ src/server/game/Entities/Creature/CreatureData.h
@@ -395,7 +395,12 @@ struct CreatureModelInfo
 };
 
 // Benchmarked: Faster than std::map (insert/find)
-typedef std::unordered_map<uint16, CreatureModelInfo> CreatureModelContainer;
+// Keyed by the full 32-bit display id. With a uint16 key every display above
+// 65535 aliased onto the stock display 65536 below it: GetCreatureModelInfo
+// returned an unrelated stock row (wrong bounding radius, combat reach, gender
+// and a random DisplayID_Other_Gender swap), or no row at all and the creature
+// refused to spawn. Stock data never goes that high; custom content does.
+typedef std::unordered_map<uint32, CreatureModelInfo> CreatureModelContainer;
 
 enum InhabitTypeValues
 {
